% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckMetaData.R
\name{CheckMetaData}
\alias{CheckMetaData}
\title{CheckMetaData}
\usage{
CheckMetaData(MetaData)
}
\arguments{
\item{MetaData}{\code{list} - A \code{list} of \code{data.frames} describing the structural and semantic characteristics of a hierarchical data model.
\itemize{ \item \code{Meta.Tables}
\item \code{Meta.Features}
\item \code{Meta.Relationships}
\item \code{Meta.Links}
\item \code{Meta.Values}}}
}
\value{
\code{list} of messages
}
\description{
Check for correctness of meta data describing a hierarchical relational data set.
}
\author{
Bastian Reiter
}
