% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanTable.R
\name{CleanTable}
\alias{CleanTable}
\title{CleanTable}
\usage{
CleanTable(
  Table,
  TableName = NA_character_,
  PrimaryKey,
  RootSubjectKey,
  SeedSubjectKey,
  PrimaryKeyIgnoredInRedundancyCheck = TRUE,
  DataSetRoot = NULL,
  UnlinkedRecords.Detect = TRUE,
  UnlinkedRecords.Remove = TRUE,
  EmptyStrings.Detect = TRUE,
  EmptyStrings.Substitute = TRUE,
  EmptyStrings.Substitution = "NA",
  DuplicateRecords.Detect = TRUE,
  DuplicateRecords.Remove = TRUE,
  ValueAvailability = NULL,
  ValueAvailabilityViolations.Detect = TRUE,
  ValueAvailabilityViolations.Remove = TRUE,
  PrintMessages = TRUE
)
}
\arguments{
\item{Table}{\code{data.frame} or \code{tibble} - The table object to be cleaned}

\item{TableName}{\code{string} - The table's name, used for command line messaging}

\item{PrimaryKey}{\code{character vector} - Name of feature(s) that serve(s) as table's primary key}

\item{RootSubjectKey}{\code{character vector} - Names of features that identify Root subjects in current table, functioning as a foreign key (usually primary key of data set Root subjects)}

\item{SeedSubjectKey}{\code{string} - The name of the feature identifying Seed subjects in table}

\item{PrimaryKeyIgnoredInRedundancyCheck}{\code{logical} - Indicating whether primary key feature has no semantic meaning and can be ignored when determining redundancy - Default: \code{TRUE}}

\item{DataSetRoot}{\code{data.frame} (Optional) - Identifying all data set Root subjects (e.g. pairs of PatientIDs and DiagnosisIDs). The \code{data.frame}'s feature names must contain foreign key features in depending tables.}

\item{UnlinkedRecords.Detect}{\code{logical}}

\item{UnlinkedRecords.Remove}{\code{logical}}

\item{EmptyStrings.Detect}{\code{logical} - Whether empty strings ('') and strings containing only white space should be detected - Default: \code{TRUE}}

\item{EmptyStrings.Substitute}{\code{logical} - Whether empty strings ('') and strings containing only white space should be removed and replaced by \code{NA} - Default: \code{TRUE}}

\item{EmptyStrings.Substitution}{\code{string} - The string which should be used to substitute empty strings. If 'NA' is passed, the empty strings are removed and set \code{NA}. Default: 'NA'}

\item{DuplicateRecords.Detect}{\code{logical} - Whether duplicate records should be detected - Default: \code{TRUE}}

\item{DuplicateRecords.Remove}{\code{logical} - Whether duplicate records should be removed - Default: \code{TRUE}}

\item{ValueAvailability}{\code{data.frame}}

\item{ValueAvailabilityViolations.Detect}{\code{logical}}

\item{ValueAvailabilityViolations.Remove}{\code{logical}}

\item{PrintMessages}{\code{logical} - Whether to print report messages during function proceedings}
}
\value{
A \code{list} containing
\itemize{ \item Table (\code{data.frame})
\item NonconformingRecords (\code{data.frame})
\item Counter (\code{data.frame})
\item Log (\code{data.frame}) }
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Detect and/or remove invalid table records.
}
\author{
Bastian Reiter
}
