% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileDictionaries.R
\name{CompileDictionaries}
\alias{CompileDictionaries}
\title{CompileDictionaries}
\usage{
CompileDictionaries(Dictionary, Dictionary.Profile, FeatureNames)
}
\arguments{
\item{Dictionary}{\code{data.frame} - Contains look-up and replacement values}

\item{Dictionary.Profile}{\code{string} - Profile selected in \emph{Dictionary}}

\item{FeatureNames}{\code{character} - Names of features to be transformed}
}
\value{
\code{list} of named vectors (serving as dictionaries)
}
\description{
Auxiliary function within \code{\link{HarmonizeData}} to compile dictionaries as named vectors.
}
\details{
Based on a given meta data.frame, compile dictionaries (named vectors) for direct value replacement
}
\author{
Bastian Reiter
}
