% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileTransFeatureRequirements.R
\name{CompileTransFeatureRequirements}
\alias{CompileTransFeatureRequirements}
\title{CompileTransFeatureRequirements}
\usage{
CompileTransFeatureRequirements(PseudoCodeRules)
}
\arguments{
\item{PseudoCodeRules}{\code{character} vector containing pseudo-code rules about feature requirements}
}
\value{
\code{character vector} containing unevaluated expressions to be used in \code{dplyr::mutate()} calls
}
\description{
Auxiliary function to compile unevaluated R expression from trans-feature requirement rules stated as pseudo-code
}
\details{
Based on a predefined rule set data.frame, compile a dplyr::case_when() expression and return it as a string
}
\author{
Bastian Reiter
}
