% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsFreda-package.R
\docType{package}
\name{dsFreda-package}
\alias{dsFreda}
\alias{dsFreda-package}
\title{dsFreda: Federated Refinement, Exploration and Data Augmentation}
\description{
Server-side DataSHIELD R package as part of a Framework for Federated Data Processing
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BastianReiter/dsFreda}
}

}
\author{
\strong{Maintainer}: Bastian Reiter \email{b.reiter@med.uni-frankfurt.de} (\href{https://orcid.org/0000-0002-1088-0502}{ORCID})

}
\keyword{internal}
