% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalizeDataHarmonization.R
\name{FinalizeDataHarmonization}
\alias{FinalizeDataHarmonization}
\title{FinalizeDataHarmonization}
\usage{
FinalizeDataHarmonization(
  TargetVector,
  EligibleValueSet,
  UnharmonizedValues.Substitute = TRUE,
  UnharmonizedValues.Substitution = ".Ineligible",
  ConvertToFactor = FALSE,
  AssignFactorLabels = FALSE
)
}
\arguments{
\item{TargetVector}{\code{vector} - Feature to be finalized}

\item{EligibleValueSet}{\code{data.frame} - Containing set of eligible values}

\item{UnharmonizedValues.Substitute}{\code{logical} - Whether to substitute all unharmonized (ineligible) values \code{NA} - Default: \code{FALSE}}

\item{UnharmonizedValues.Substitution}{\code{string} - Which string should substitute unharmonized (ineligible) values? If set "NA", the values will be substituted by \code{NA}}

\item{ConvertToFactor}{\code{logical} - Whether to convert vector to factor - Default: \code{FALSE}}

\item{AssignFactorLabels}{\code{logical} - Whether to assign factor labels during factor conversion - Default: \code{FALSE}}
}
\value{
A \code{vector} (or factor)
}
\description{
Auxiliary function within \code{\link{CurateDataDS}}
}
\details{
Ineligible data (including data that could not be transformed) is turned into NA per default.
Optional factor conversion establishes level order where appropriate.
}
\author{
Bastian Reiter
}
