% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalizeDataRemediation.R
\name{FinalizeDataRemediation}
\alias{FinalizeDataRemediation}
\title{FinalizeDataRemediation}
\usage{
FinalizeDataRemediation(
  TargetVector,
  EligibleValueSet,
  UnremediatedValues.Substitute = TRUE,
  UnremediatedValues.Substitution = ".Ineligible",
  ConvertToFactor = FALSE,
  AssignFactorLabels = FALSE
)
}
\arguments{
\item{TargetVector}{\code{vector} - Feature to be finalized}

\item{EligibleValueSet}{\code{data.frame} - Containing set of eligible values}

\item{UnremediatedValues.Substitute}{\code{logical} - Whether to substitute all unremediated (ineligible) values \code{NA} - Default: \code{TRUEs}}

\item{UnremediatedValues.Substitution}{\code{string} - Which string should substitute unremediated (ineligible) values? If set "NA", the values will be substituted by \code{NA}}

\item{ConvertToFactor}{\code{logical} - Whether to convert vector to factor - Default: \code{FALSE}}

\item{AssignFactorLabels}{\code{logical} - Whether to assign factor labels during factor conversion - Default: \code{FALSE}}
}
\value{
A \code{vector} (or factor)
}
\description{
Auxiliary function within \code{\link{CurateDataDS}}
}
\details{
Unremediated / Ineligible data is turned into NA per default.
Optional factor conversion establishes level order where appropriate.
}
\author{
Bastian Reiter
}
