% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRedundantRecords.R
\name{FindRedundantRecords}
\alias{FindRedundantRecords}
\title{FindRedundantRecords}
\usage{
FindRedundantRecords(
  Table,
  PrimaryKeyFeature,
  DiscriminatoryFeatures,
  EssentialFeatures,
  RemoveRedundantRecords = FALSE
)
}
\arguments{
\item{Table}{\code{data.frame} or \code{tibble}}

\item{PrimaryKeyFeature}{\code{string} - Name of primary key feature}

\item{DiscriminatoryFeatures}{\code{character} - Names of features that are used to strictly discriminate between different table records (used in \code{group_by}-statement)}

\item{EssentialFeatures}{\code{character} - Names of features that are considered essential in the informational value of a table record}

\item{RemoveRedundantRecords}{\code{logical} - Indicates whether to remove table records that are considered redundant. If set to \code{FALSE}, the redundant records are marked as such and preserved. - Default: \code{FALSE}}
}
\value{
\code{data.frame} or \code{tibble} cleaned of redundant records
}
\description{
Auxiliary function within \code{\link{CurateDataDS}}
}
\details{
Identify and remove any records that contain less essential information compared to a previous record.
}
\author{
Bastian Reiter
}
