% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindSubsumptionRedundancies.R
\name{FindSubsumptionRedundancies}
\alias{FindSubsumptionRedundancies}
\title{FindSubsumptionRedundancies}
\usage{
FindSubsumptionRedundancies(
  Table,
  PrimaryKey,
  DistinctiveFeatures,
  NegligibleFeatures = NULL,
  NegligibleValues = NULL
)
}
\arguments{
\item{Table}{\code{data.frame} or \code{tibble}}

\item{PrimaryKey}{\code{character vector} - Name of feature(s) that serve(s) as table's primary key}

\item{DistinctiveFeatures}{\code{character vector} - Names of features that are used to strictly distinguish different table records (used in \code{group_by}-statement)}

\item{NegligibleFeatures}{\code{character vector} - Names of features that are considered negligible in the informational value of a table record}

\item{NegligibleValues}{\code{list} - Which feature values do not have informational value? List with element names being feature names and list elements being character vectors with negligible values of the feature.}
}
\value{
The input Table with records being marked as redundant by subsumption
}
\description{
Auxiliary function within \code{\link{SubsumeRecords}}
}
\details{
Identify and remove any records that contain less essential information compared to a previous record.
}
\author{
Bastian Reiter
}
