% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataSetCheckDS.R
\name{GetDataSetCheckDS}
\alias{GetDataSetCheckDS}
\title{GetDataSetCheckDS}
\usage{
GetDataSetCheckDS(
  DataSetName.S,
  RequiredTableNames.S = NULL,
  RequiredFeatureNames.S = NULL,
  EligibleValueSets.S = NULL,
  Module.S = NULL,
  Stage.S = NULL
)
}
\arguments{
\item{DataSetName.S}{\code{string} - Name of Data Set object (list) on server, usually "RawDataSet", "CuratedDataSet" or "AugmentedDataSet"}

\item{RequiredTableNames.S}{Optional \code{character vector} - Names of required tables}

\item{RequiredFeatureNames.S}{Optional \code{list} of \code{character vectors} - Names of required features - Default: \code{names(Table)}}

\item{EligibleValueSets.S}{Optional \code{list} of \code{character vectors} containing sets of eligible values for corresponding feature.}

\item{Module.S}{Optional \code{string} identifying a defined data set and the corresponding meta data (Examples: 'CCP' / 'P21')}

\item{Stage.S}{Optional \code{string} - Indicating transformation stage of addressed data set. This is relevant for which names and values to look up in module meta data. Options: 'Raw' / 'Curated' / 'Augmented'}
}
\value{
A \code{list} containing meta data about tables in a data set
}
\description{
Checks out a data set (\code{list} of \code{data.frames}) and returns an informative list object. Options:
\enumerate{
\item The data set can be inspected naively without passing any additional information about it
\item Meta data like required table and feature names as well as eligible value sets can be passed explicitly
\item A Module identifier from a list of registered modules can be passed ('CCP' / 'P21' / ...), which leads to meta data being taken from a linked package }
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
