% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEcdfDS.R
\name{GetEcdfDS}
\alias{GetEcdfDS}
\title{GetEcdfDS}
\usage{
GetEcdfDS(
  TableName.S,
  FeatureName.S,
  GroupingFeatureName.S = NULL,
  RemoveNA.S = TRUE
)
}
\arguments{
\item{TableName.S}{\code{string} - Name of the \code{data.frame} holding relevant feature(s)}

\item{FeatureName.S}{\code{string} - Name of numeric feature}

\item{GroupingFeatureName.S}{\code{string} - Name of optional grouping feature}
}
\value{
A function of class \code{ecdf}, computed by \code{stats::ecdf()}
}
\description{
Return empirical cumulative distribution function for a numeric feature.
}
