% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetErrorTraceDS.R
\name{GetErrorTraceDS}
\alias{GetErrorTraceDS}
\title{GetErrorTraceDS}
\usage{
GetErrorTraceDS(FunctionCall.S)
}
\arguments{
\item{FunctionCall.S}{\code{call}}
}
\value{
A
}
\description{
Run a server-side function while tracing calls. In an error occurs, the trace is returned to the client.
}
\details{
Server-side AGGREGATE method
}
