% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFrequencyTableDS.R
\name{GetFrequencyTableDS}
\alias{GetFrequencyTableDS}
\title{GetFrequencyTableDS}
\usage{
GetFrequencyTableDS(
  TableName.S,
  FeatureName.S,
  GroupingFeatureName.S = NULL,
  RemoveNA.S = FALSE
)
}
\arguments{
\item{TableName.S}{\code{string} - Name of the Data frame that contains the feature}

\item{FeatureName.S}{\code{string} - Name of feature}

\item{GroupingFeatureName.S}{\code{string} - Name of optional grouping feature}

\item{RemoveNA.S}{\code{logical} - Indicating whether missing values should be removed prior to frequency calculation - Default: \code{FALSE}}
}
\value{
A \code{tibble} containing absolute and relative frequencies
}
\description{
Return table of absolute and relative value frequencies for a nominal / ordinal feature.
}
\author{
Bastian Reiter
}
