% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetObjectStatusDS.R
\name{GetObjectStatusDS}
\alias{GetObjectStatusDS}
\title{GetObjectStatusDS}
\usage{
GetObjectStatusDS(ObjectName.S)
}
\arguments{
\item{ObjectName.S}{\code{string} - Name of object on server}
}
\value{
A \code{list} containing information about existence and class of object
}
\description{
Similar to \code{dsBase::testObjExistsDS()}. Tests whether a certain object exists physically on server and provides meta info about it.
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
