% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSampleStatisticsDS.R
\name{GetSampleStatisticsDS}
\alias{GetSampleStatisticsDS}
\title{GetSampleStatisticsDS}
\usage{
GetSampleStatisticsDS(
  TableName.S,
  FeatureName.S,
  GroupingFeatureName.S = NULL,
  RemoveNA.S = TRUE,
  ReturnECDF.S = TRUE
)
}
\arguments{
\item{TableName.S}{\code{string} - Name of the table that contains the feature}

\item{FeatureName.S}{\code{string} - Name of feature}

\item{GroupingFeatureName.S}{\code{string} - Name of optional grouping feature}

\item{RemoveNA.S}{\code{logical} - Whether NA values should be removed before calculating statistics}

\item{ReturnECDF.S}{\code{logical} - Whether an empirical cumulative distribution function should be returned additionally to sample statistics}
}
\value{
A \code{tibble} containing parametric and non-parametric sample statistics
}
\description{
Calculate common sample statistics for a feature of a data.frame.
}
\author{
Bastian Reiter
}
