% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSystemSpecsDS.R
\name{GetSystemSpecsDS}
\alias{GetSystemSpecsDS}
\title{GetSystemSpecsDS}
\usage{
GetSystemSpecsDS()
}
\value{
A \code{list} containing
\itemize{ \item CPU.Count - Number of physical or logical CPU cores
\item RAM.Total - Total installed RAM
\item RAM.Available - RAM that is available instantly }
}
\description{
Get CPU and RAM specifications of local system using \code{ps} package functions
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
