% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTestDataDS.R
\name{GetTestDataDS}
\alias{GetTestDataDS}
\title{GetTestDataDS}
\usage{
GetTestDataDS(
  DataSetName.S = NA_character_,
  TableName.S = NA_character_,
  SampleSize.S,
  Shuffle.S = TRUE
)
}
\arguments{
\item{DataSetName.S}{\code{string} - Name of a \code{list} of \code{data.frames} representing a data set}

\item{TableName.S}{\code{string} - Name of a specific \code{data.frame}}
}
\value{
A
}
\description{
Get randomized test data from servers (only possible in 'developer' mode).
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
