% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizeFeature.R
\name{HarmonizeFeature}
\alias{HarmonizeFeature}
\title{HarmonizeFeature}
\usage{
HarmonizeFeature(
  Vector,
  Methods,
  EligibleValueSet = NULL,
  TransformativeExpressions = NULL,
  Dictionary = NULL,
  FuzzyStringMatching = NULL
)
}
\arguments{
\item{Vector}{\code{vector} containing data to be transformed}

\item{Methods}{\code{list} - Contains data on the selection of methods to use for each feature}

\item{EligibleValueSet}{\code{character} vector containing set of eligible feature values}

\item{TransformativeExpressions}{\code{data.frame} - Contains set of expressions like functions used to transform data values}

\item{Dictionary}{\code{character} - Contains dictionary data used to look up and replace data values}

\item{FuzzyStringMatching}{\code{list} - Contains settings for Fuzzy String Matching}
}
\value{
The input \code{data.frame} with transformed data values
}
\description{
This function performs harmonizing data transformation for features of a \code{data.frame} based on passed settings.
Harmonization is attempted in a step-wise approach, incorporating various methods in a defined order:
\enumerate{\item Transformative expressions (e.g. functions like \code{str_to_upper()})
\item Dictionary look-up
\item Fuzzy String Matching}
}
\author{
Bastian Reiter
}
