% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeListDS.R
\name{MakeListDS}
\alias{MakeListDS}
\title{MakeListDS}
\usage{
MakeListDS(ObjectNames.S)
}
\arguments{
\item{ObjectNames.S}{\code{character} - Names of objects to be bundled in a list. Can be a \code{named vector} if names in list should differ from symbol names (with names in named vector being the names in list).}
}
\value{
A \code{data.frame} resulting from filter operation
}
\description{
Bundle objects in a \code{list} object.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
