% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta.Modules}
\alias{Meta.Modules}
\title{Registered modules ('CCP' / 'P21' / ...)}
\format{
\subsection{\code{Meta.Modules}}{

\code{list}
}
}
\usage{
Meta.Modules
}
\description{
A \code{list} containing names of modules and linked package names
}
\keyword{datasets}
