% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ModuleRegistration}
\alias{ModuleRegistration}
\title{Registered modules ('CCP' / 'P21' / ...)}
\format{
\subsection{\code{ModuleRegistration}}{

\code{list}
}
}
\usage{
ModuleRegistration
}
\description{
A \code{list} containing names of modules and linked package names
}
\keyword{datasets}
