% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeTable.R
\name{NormalizeTable}
\alias{NormalizeTable}
\title{NormalizeTable}
\usage{
NormalizeTable(
  Table,
  TableName,
  PrimaryKey,
  ForeignKey = NULL,
  RuleSet,
  RuleSet.Profile
)
}
\arguments{
\item{Table}{\code{data.frame} containing data to be transformed}

\item{TableName}{\code{string} - Name of the table (to enable mapping to normalization rules)}

\item{PrimaryKey}{\code{character vector} - Name of features that serve as table's primary key}

\item{ForeignKey}{\code{character vector} - Names of features that serve as table's foreign key (usually primary key of data set 'root subjects')}

\item{RuleSet}{\code{data.frame} - Contains predefined set of normalization rules}

\item{RuleSet.Profile}{\code{string} - Profile name stated in 'RuleSet'}
}
\value{
The transformed input \code{data.frame}
}
\description{
Carries out table harmonization procedures (like 'splitting/expanding') based on a given rule set.
}
\details{
Uses \code{tidyr} functionality to transform table based on normalization rules defined in a given rule set (Default: Proc.TableNormalization)
}
\author{
Bastian Reiter
}
