% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareRawDataDS.R
\name{PrepareRawDataDS}
\alias{PrepareRawDataDS}
\title{PrepareRawDataDS}
\usage{
PrepareRawDataDS(
  RawDataSetName.S,
  Module.S,
  FeatureNames.Dictionary.S = list(),
  FeatureNames.FuzzyStringMatching.Run.S = FALSE,
  FeatureNames.FuzzyStringMatching.PreferredMethod.S = "jw",
  FeatureNames.FuzzyStringMatching.Tolerance.S = 0.2,
  AddIDFeature.Do.S = FALSE,
  AddIDFeature.IDFeatureName.S = "ID",
  AddIDFeature.OverwriteExistingIDFeature.S = FALSE,
  Conversion.IntoCharacter.S = "None",
  Conversion.DateIntoPOSIXct.S = NULL,
  CurateFeatureNames.S = FALSE
)
}
\arguments{
\item{RawDataSetName.S}{\code{string} - Name of Raw Data Set object (list) on server - Default: 'P21.RawDataSet'}

\item{Module.S}{\code{string} identifying a defined data set and the corresponding meta data needed for feature name harmonization (Examples: 'CCP' / 'P21')}

\item{FeatureNames.Dictionary.S}{Optional \code{list} containing dictionary data for raw feature name harmonization (Form: \code{list(Department = c(FAB = "Fachabteilung"))})}

\item{FeatureNames.FuzzyStringMatching.Run.S}{\code{logical} - Whether to use fuzzy string matching to harmonize raw feature names}

\item{FeatureNames.FuzzyStringMatching.PreferredMethod.S}{\code{logical} - Selects the method \code{stringdist()} uses (see \code{stringdist} documentation) - Default: 'jw'}

\item{FeatureNames.FuzzyStringMatching.Tolerance.S}{\code{logical} - Number between 0 and 1 relating to normalized distance between to strings (1 meaning furthest distance, 0 meaning no distance). If a string in 'Vector' is not similar enough to any of the 'EligibleStrings' and its minimal harmonized distance exceeds this number, it is set \code{NA}. - Default: 0.2}

\item{AddIDFeature.Do.S}{\code{logical} - Whether to add an ID feature (running number) to tables}

\item{AddIDFeature.IDFeatureName.S}{\code{string} - The name of the new ID feature}

\item{AddIDFeature.OverwriteExistingIDFeature.S}{\code{logical} - Whether to overwrite an existing feature with the same name}

\item{Conversion.IntoCharacter.S}{\code{string} - Controls conversion of certain features in data set tables into character type. One of 'None' / 'All' / 'Date'. - Default: 'None'}

\item{Conversion.DateIntoPOSIXct.S}{\code{string} - An encoded string representing a list with character vectors containing date formats used in argument 'tryFormats' in function \code{base::as.POSIXct}. List element should be either '.All' for all date features or specific date feature names. - Default: \code{NULL}}

\item{CurateFeatureNames.S}{\code{logical} - Indicating whether (after primary harmonization) feature names should be recoded from 'raw' to 'curated' feature names according to Module-specific meta data}
}
\value{
A \code{list} containing
\itemize{ \item The input RawDataSet (\code{list}) with harmonized (and optionally curated) feature names
\item The original input RawDataSet (\code{list})
\item Messages (\code{character vector}) }
}
\description{
Perform basic preparatory transformations on RawDataSet prior to Curation:
\itemize{  \item Feature name harmonization
\item Add ID feature (running number) for tables without primary key feature
\item Type- or format-conversion of features }
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
