% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.DataHarmonizationMethods}
\alias{Set.DataHarmonizationMethods}
\title{Settings on which methods to use for each feature during data harmonization}
\format{
\subsection{\code{Set.DataHarmonizationMethods}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{FeatureID}{}
\item{Table}{}
\item{Feature}{}
\item{HasEligibleValueSet}{}
\item{TransformationRules}{}
\item{HashTable}{}
\item{FuzzyStringMatching}{}
\item{Classification}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.DataHarmonizationMethods
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
