% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.PrivacyProfile.Chosen}
\alias{Set.PrivacyProfile.Chosen}
\title{Chosen privacy profile}
\format{
\subsection{\code{Set.PrivacyProfile.Chosen}}{

\code{string}
}
}
}
\usage{
Set.PrivacyProfile.Chosen
}
\description{
Chosen privacy profile
}
\keyword{datasets}
