% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.PrivacyProfiles}
\alias{Set.PrivacyProfiles}
\title{Different profiles of privacy settings used to determine disclosure control in relevant dsFreda functions}
\format{
\subsection{\code{Set.PrivacyProfiles}}{

\code{list}
\describe{
\item{Profile}{'Development' / 'Loose' or 'Strict'}
\item{NThreshold}{The minimum sample size required for transmission of aggregated data to client}}
}
}
\usage{
Set.PrivacyProfiles
}
\description{
Different profiles of privacy settings used to determine disclosure control in relevant dsFreda functions
}
\keyword{datasets}
