% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackValueCounts.R
\name{TrackValueCounts}
\alias{TrackValueCounts}
\title{TrackValueCounts}
\usage{
TrackValueCounts(
  DataFrame,
  FeatureNames,
  EligibleValues = NULL,
  TransformationStage = NULL
)
}
\arguments{
\item{DataFrame}{\code{data.frame}}

\item{FeatureNames}{\code{character} vector containing names of features to be tracked}

\item{EligibleValues}{An optional \code{list} of vectors containing eligible value sets. Name of vector object gives name of feature, vector values give eligible values.}

\item{TransformationStage}{String representing transformation stage (usually "Raw" / "Transformed" / "Final")}
}
\value{
A \code{tibble}
}
\description{
Track occurrence of a feature's unique values, their eligibility and frequency
}
\author{
Bastian Reiter
}
