% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanTable.R
\name{CleanTable}
\alias{CleanTable}
\title{CleanTable}
\usage{
CleanTable(
  Table,
  TableNameLookup,
  PrimaryKey,
  ForeignKey = NULL,
  PrimaryKeyIgnoredInRedundancyCheck = TRUE,
  RemoveEmptyStrings = TRUE,
  RemoveDuplicateEntries = TRUE,
  FeatureObligations = NULL,
  ObeyFeatureObligations = TRUE,
  ObeyTransFeatureObligations = TRUE
)
}
\arguments{
\item{Table}{\code{data.frame} or \code{tibble} - The table object to be cleaned}

\item{TableNameLookup}{\code{string} - The table name(s) to be looked up in 'FeatureObligations$RuleSet'. Can be a single string or a character vector.}

\item{PrimaryKey}{\code{character vector} - Name of features that serve as primary key for table}

\item{ForeignKey}{\code{character vector} - Names of features that serve as foreign key for table (usually primary key of data set 'root subjects')}

\item{PrimaryKeyIgnoredInRedundancyCheck}{\code{logical} - Indicating whether primary key feature has no semantic meaning and can be ignored when determining redundancy - Default: \code{TRUE}}

\item{RemoveEmptyStrings}{\code{logical} - Whether empty strings ('') and strings containing only white space should be removed and replaced by \code{NA} - Default: \code{TRUE}}

\item{RemoveDuplicateEntries}{\code{logical} - Whether duplicate entries should be removed}

\item{FeatureObligations}{\code{list}
\itemize{\item \code{RuleSet} - \code{data.frame}
\item \code{RuleSet.Profile} - \code{character} - Profile name defining strict and trans-feature rules for obligatory feature content. Profile name must be stated in \code{FeatureObligations$RuleSet}}}

\item{ObeyFeatureObligations}{\code{logical}}

\item{ObeyTransFeatureObligations}{\code{logical}}
}
\value{
\code{tibble} - Clean table
}
\description{
Perform exclusion of invalid table entries.
}
\author{
Bastian Reiter
}
