% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurateDataDS.R
\name{CurateDataDS}
\alias{CurateDataDS}
\title{CurateDataDS}
\usage{
CurateDataDS(
  RawDataSetName.S = "RawDataSet",
  Module.S,
  Profile.CurationProcess.S = "Default",
  Profile.DataRemediation.S = "Default",
  Profile.Dictionary.S = "Default",
  Profile.FeatureRequirements.S = "Default",
  Profile.FeatureTracking.S = "Default",
  Profile.FuzzyStringMatching.S = "Default",
  Profile.PrimaryTableCleaning.S = "Default",
  Profile.RecordSubsumption.S = "Default",
  Profile.SecondaryTableCleaning.S = "Default",
  Profile.TableNormalization.S = "Default",
  Profile.TransformativeExpressions.S = "Default"
)
}
\arguments{
\item{RawDataSetName.S}{\code{character} - Name of Raw Data Set object (list) on server - Default: 'RawDataSet'}

\item{Module.S}{\code{string} - Identifying a registered FREDA module. Meta data and processing settings are obtained from the corresponding installed package.}

\item{Profile.CurationProcess.S}{\code{string} - "Default"}

\item{Profile.DataRemediation.S}{\code{string} - "Default"}

\item{Profile.Dictionary.S}{\code{string} - "Default"}

\item{Profile.FeatureRequirements.S}{\code{string} - "Default"}

\item{Profile.FeatureTracking.S}{\code{string} - "Default"}

\item{Profile.FuzzyStringMatching.S}{\code{string} - "Default"}

\item{Profile.PrimaryTableCleaning.S}{\code{string} - "Default"}

\item{Profile.RecordSubsumption.S}{\code{string} - "Default"}

\item{Profile.SecondaryTableCleaning.S}{\code{string} - "Default"}

\item{Profile.TableNormalization.S}{\code{string} - "Default"}

\item{Profile.TransformativeExpressions.S}{\code{string} - "Default"}
}
\value{
A \code{list} containing the following objects:
\itemize{\item DataSet \code{list}
\item NonconformingRecords \code{list}
\item AffectedRootSubjects \code{list}
\itemize{ \item HadNonconformingRecords
\item HadRemovedRecords }
\item Report \code{list}
\itemize{ \item Settings
\item Process
\item Log
\item Counter
\itemize{ \item Summary
\itemize{ \item DataSetLevel
\item TableLevel }
\item Extensive
\itemize{ \item StageLevel
\item Details }}
\item DataHarmonization
\itemize{ \item Reports
\itemize{ \item DataSetLevel
\item TableLevel
\item FeatureLevel
\item ValueLevel
\item ValueSets }
\item Monitors
\itemize{ \item TableLevel
\item FeatureLevel
\item ValueLevel }}}
\item Messages \code{list}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Transforms Raw Data Set (RDS) into Curated Data Set (CDS) while tracking data set transformation.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
