% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurateDataDS.R
\name{CurateDataDS}
\alias{CurateDataDS}
\title{CurateDataDS}
\usage{
CurateDataDS(
  RawDataSetName.S = "RawDataSet",
  MetaData.S = list(Tables = dsCCPhos::Meta.Tables, Features = dsCCPhos::Meta.Features,
    Values = dsCCPhos::Meta.Values),
  Proc.TableNormalization.S = dsCCPhos::Proc.TableNormalization,
  Proc.TableNormalization.Profile.S = "Default",
  Set.CurationProcess.S = dsCCPhos::Set.CurationProcess,
  Set.CurationProcess.Profile.S = "Default",
  Set.DataHarmonization.S = dsCCPhos::Set.DataHarmonization,
  Set.DataHarmonization.Profile.S = "Default",
  Set.DataHarmonization.TransformativeExpressions.S =
    dsCCPhos::Set.TransformativeExpressions,
  Set.DataHarmonization.TransformativeExpressions.Profile.S = "Default",
  Set.DataHarmonization.Dictionary.S = dsCCPhos::Set.Dictionary,
  Set.DataHarmonization.Dictionary.Profile.S = "Default",
  Set.DataHarmonization.FuzzyStringMatching.S = dsCCPhos::Set.FuzzyStringMatching,
  Set.DataHarmonization.FuzzyStringMatching.Profile.S = "Default",
  Set.FeatureRequirements.S = dsCCPhos::Set.FeatureRequirements,
  Set.FeatureRequirements.Profile.S = "Default",
  Set.FeatureTracking.S = dsCCPhos::Set.FeatureTracking,
  Set.FeatureTracking.Profile.S = "Default",
  Set.PrimaryTableCleaning.S = dsCCPhos::Set.PrimaryTableCleaning,
  Set.PrimaryTableCleaning.Profile.S = "Default",
  Set.SecondaryTableCleaning.S = dsCCPhos::Set.SecondaryTableCleaning,
  Set.SecondaryTableCleaning.Profile.S = "Default",
  Set.SubsumptionRedundancies.S = dsCCPhos::Set.SubsumptionRedundancies,
  Set.SubsumptionRedundancies.Profile.S = "Default"
)
}
\arguments{
\item{RawDataSetName.S}{\code{character} - Name of Raw Data Set object (list) on server - Default: 'RawDataSet'}

\item{MetaData.S}{\code{list} - A \code{list} of \code{data.frames} describing the structural and semantic characteristics of a hierarchical data model.
\itemize{ \item \code{Meta.Tables}
\item \code{Meta.Features}
\item \code{Meta.Values}}}

\item{Proc.TableNormalization.S}{\code{data.frame} - Default: \code{dsCCPhos::Proc.TableNormalization}}

\item{Proc.TableNormalization.Profile.S}{\code{string} - Profile name defining rule set to be used for table normalization. Profile name must be stated in \code{TableNormalization$RuleSet} - Default: 'Default'}

\item{Set.CurationProcess.S}{\code{data.frame}}

\item{Set.CurationProcess.Profile.S}{\code{string}}

\item{Set.DataHarmonization.S}{\code{data.frame} - Default: \code{dsCCPhos::Set.DataHarmonization}}

\item{Set.DataHarmonization.Profile.S}{\code{string} - Profile used in \emph{DataHarmonization} - Default: 'Default'}

\item{Set.DataHarmonization.TransformativeExpressions.S}{\code{data.frame} - Default: \code{dsCCPhos::Set.TransformativeExpressions}}

\item{Set.DataHarmonization.TransformativeExpressions.Profile.S}{\code{string} - Profile used in \emph{TransformativeExpressions} - Default: 'Default'}

\item{Set.DataHarmonization.Dictionary.S}{\code{data.frame} - Default: \code{dsCCPhos::Set.Dictionary}}

\item{Set.DataHarmonization.Dictionary.Profile.S}{\code{string} - Profile used in \emph{Dictionary} - Default: 'Default'}

\item{Set.DataHarmonization.FuzzyStringMatching.S}{\code{data.frame} - Default: \code{dsCCPhos::Set.FuzzyStringMatching}}

\item{Set.DataHarmonization.FuzzyStringMatching.Profile.S}{\code{string} - Profile used in \emph{FuzzyStringMatching} - Default: 'Default'}

\item{Set.FeatureRequirements.S}{\code{data.frame} - Default: \code{dsCCPhos::Set.FeatureRequirements}}

\item{Set.FeatureRequirements.Profile.S}{\code{string} - Profile name defining strict and trans-feature requirements - Default: 'Default'}

\item{Set.FeatureTracking.S}{\code{data.frame} - Default: \code{dsCCPhos::Set.FeatureTracking}}

\item{Set.FeatureTracking.Profile.S}{\code{string} - Profile name defining which features should be tracked/monitored during curation process. Profile name must be stated in \code{FeatureTracking$RuleSet} - Default: 'Default'}

\item{Set.PrimaryTableCleaning.S}{\code{data.frame} - Default: \code{dsCCPhos::Set.PrimaryTableCleaning}}

\item{Set.PrimaryTableCleaning.Profile.S}{\code{string} - Default: 'Default'}

\item{Set.SecondaryTableCleaning.S}{\code{data.frame} - Default: \code{dsCCPhos::Set.SecondaryTableCleaning}}

\item{Set.SecondaryTableCleaning.Profile.S}{\code{string} - Default: 'Default'}

\item{Set.SubsumptionRedundancies.S}{\code{data.frame} - Default: \code{dsCCPhos::Set.SubsumptionRedundancies}}

\item{Set.SubsumptionRedundancies.Profile.S}{\code{string} - Default: 'Default'}
}
\value{
A \code{list} containing the following objects:
\itemize{\item CuratedDataSet \code{list}
\itemize{ \item BioSampling
\item Diagnosis
\item DiseaseStatus
\item GeneralCondition
\item Histology
\item Metastasis
\item MolecularDiagnostics
\item OtherClassification
\item Patient
\item RadiationTherapy
\item Staging
\item Surgery
\item SystemicTherapy
\item TherapyRecommendation}
\item CurationReport \code{list}
\itemize{ \item RecordCounts \code{tibble}
\itemize{ \item Table
\item InitialCount
\item ExcludedPrimary
\item AfterPrimaryExclusion
\item ExcludedSecondary
\item AfterSecondaryExclusion
\item ExcludedSecondaryRedundancy
\item AfterSecondaryRedundancyExclusion}
\item Transformation (list of lists)
\itemize{ \item Monitors
\item EligibilityOverviews
\item ValueSetOverviews}}
\item Messages \code{list}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Transforms Raw Data Set (RDS) into Curated Data Set (CDS) while tracking data set transformation.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
