% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataSetCheckDS.R
\name{GetDataSetCheckDS}
\alias{GetDataSetCheckDS}
\title{GetDataSetCheckDS}
\usage{
GetDataSetCheckDS(
  DataSetName.S,
  DataSetMetaData.S = NULL,
  Module.S = NULL,
  Stage.S = "Raw"
)
}
\arguments{
\item{DataSetName.S}{\code{string} - Name of Data Set object (list) on server, usually "RawDataSet", "CuratedDataSet" or "AugmentedDataSet"}

\item{DataSetMetaData.S}{Optional \code{list} of \code{data.frames} 'Meta.Tables' / 'Meta.Features' / 'Meta.Values'}

\item{Module.S}{Optional \code{string} identifying a defined data set (Examples: 'CCP' / 'P21')}

\item{Stage.S}{Optional \code{string} - Indicating transformation stage of addressed data set. This is relevant for which names and values to look up in passed meta data. Options: 'Raw' / 'Curated'}
}
\value{
A \code{list} containing meta data about tables in a data set
}
\description{
Checks out a data set (list of data.frames) and returns an informative list object. Options:
\enumerate{
\item The data set can be inspected naively without passing any additional information about it
\item Meta data about the data set can be passed in the form of a \code{list} containing the \code{data.frames} 'Meta.Tables' / 'Meta.Features' / 'Meta.Values'
\item A Module identifier from a list of registered modules can be passed ('CCP' / 'P21' / ...), which leads to meta data being taken from a linked package }
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
