% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeTable.R
\name{NormalizeTable}
\alias{NormalizeTable}
\title{NormalizeTable}
\usage{
NormalizeTable(Table, PrimaryKey, ForeignKey = NULL, RuleSet)
}
\arguments{
\item{Table}{\code{data.frame} containing data to be transformed}

\item{PrimaryKey}{\code{character vector} - Name of features that serve as table's primary key}

\item{ForeignKey}{\code{character vector} - Names of features that serve as table's foreign key (usually primary key of data set root subjects)}

\item{RuleSet}{\code{data.frame} - Contains predefined set of normalization rules}
}
\value{
The transformed input \code{data.frame}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Carries out table harmonization procedures (like 'splitting/expanding') based on a given rule set.
}
\details{
Uses \code{tidyr} functionality to transform table based on normalization rules defined in a given rule set (Default: Proc.TableNormalization)
}
\author{
Bastian Reiter
}
