% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeTable.R
\name{NormalizeTable}
\alias{NormalizeTable}
\title{NormalizeTable}
\usage{
NormalizeTable(
  Table,
  TableName = NA_character_,
  PrimaryKey,
  RootSubjectKey = NULL,
  RuleSet,
  PrintMessages = TRUE
)
}
\arguments{
\item{Table}{\code{data.frame} containing data to be transformed}

\item{TableName}{\code{string} - The table's name, used for command line messaging}

\item{PrimaryKey}{\code{character vector} - Name of features that serve as table's primary key}

\item{RootSubjectKey}{\code{character vector} - Names of features that identify root subjects in current table, functioning as a foreign key (usually primary key of data set root subjects)}

\item{RuleSet}{\code{data.frame} - Contains predefined set of normalization rules}

\item{PrintMessages}{\code{logical} - Whether to print report messages during function proceedings}
}
\value{
A \code{list} containing:
\itemize{ \item Table (\code{data.frame})
\item Tracker (\code{data.frame})
\item Log (\code{data.frame}) }
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Carries out table normalization procedures (like 'splitting/expanding') based on a given rule set.
}
\details{
Uses \code{tidyr} functionality to transform table based on normalization rules defined in a given rule set (Default: Proc.TableNormalization)
}
\author{
Bastian Reiter
}
