% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareRawDataDS.R
\name{PrepareRawDataDS}
\alias{PrepareRawDataDS}
\title{PrepareRawDataDS}
\usage{
PrepareRawDataDS(
  RawDataSetName.S,
  Module.S,
  FeatureNameDictionary.S = list(),
  RunFuzzyStringMatching.S = FALSE,
  FSMSettings.S = list(PreferredMethod = "jw", Tolerance = 0.2),
  AddIDFeature.S = list(Do = FALSE, IDFeatureName = "ID", OverwriteExistingIDFeature =
    FALSE),
  CompleteCharacterConversion.S = FALSE,
  CurateFeatureNames.S = FALSE
)
}
\arguments{
\item{RawDataSetName.S}{\code{string} - Name of Raw Data Set object (list) on server - Default: 'P21.RawDataSet'}

\item{Module.S}{\code{string} identifying a defined data set and the corresponding meta data needed for feature name harmonization (Examples: 'CCP' / 'P21')}

\item{FeatureNameDictionary.S}{Optional \code{list} containing dictionary data for raw feature name harmonization (Form: \code{list(Department = c(FAB = "Fachabteilung"))})}

\item{RunFuzzyStringMatching.S}{\code{logical} - Whether to use fuzzy string matching to harmonize raw feature names}

\item{FSMSettings.S}{\code{list} of parameters for Fuzzy String Matching ('PreferredMethod', 'Tolerance')}

\item{AddIDFeature.S}{\code{list} containing parameters about adding an ID feature to tables:
\itemize{ \item Do (\code{logical}) - Whether to add an ID feature (running number)
\item IDFeatureName (\code{string})
\item OverwriteExistingIDFeature (\code{logical}) - Whether to overwrite an existing feature with the same name }}

\item{CompleteCharacterConversion.S}{\code{logical} - Indicating whether to convert all features in data set tables to character type}

\item{CurateFeatureNames.S}{\code{logical} - Indicating whether (after primary harmonization) feature names should be recoded from 'raw' to 'curated' feature names according to Module-specific meta data}
}
\value{
A \code{list} containing
\itemize{ \item The input RawDataSet (\code{list}) with harmonized (and optionally curated) feature names
\item The original input RawDataSet (\code{list})
\item Messages (\code{character vector}) }
}
\description{
Perform basic preparatory transformations on RawDataSet prior to Curation:
\itemize{  \item Optionally convert all columns to character type
\item Add ID feature (running number) for tables without primary key feature
\item Try to harmonize feature names employing Fuzzy String Matching and Dictionary look-up }
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
